requirements.txt

protobuf==3.20
python-decouple
tinkoff-voicekit-client
PyAudio
grpcio
openai